+++ app/javascript/users/components/homepage/homepage_sections.vue
@@ -4,7 +4,7 @@
 			<v-col cols="12">
 				<component
 						:is="`section-${section}`"
-						v-for="section in sections"
+						v-for="section in showed_sections"
 						:key="section"
 				/>
 			</v-col>
@@ -13,6 +13,8 @@
 </template>
 
 <script>
+  import { mapState } from 'vuex'
+
   export default {
     name: 'BaseView',
 
@@ -24,5 +26,19 @@
     },
 
     data: () => ({ sections: [] }),
+
+		// created() {
+    //   console.log(`BaseView sections:`, this.sections)
+		// },
+    computed: {
+      ...mapState('application', ['current_organization']),
+      showed_sections() {
+        if (this.current_organization.homepage) {
+          return this.sections.filter(x => !['affiliates', 'pro-features'].includes(x))
+        } else {
+          return this.sections
+        }
+      },
+    },
   }
 </script>