+++ app/javascript/users/mixins/homepage_content_mixin.js
@@ -1,0 +1,13 @@
+import { mapState } from 'vuex'
+
+export const homepageContentMixin = {
+  computed: {
+    ...mapState('application', ['current_organization']),
+    organizationHomepageData() {
+      return this.current_organization.homepage
+    },
+    organizationEmail() {
+      return this.current_organization.email || 'support@insiemento.com'
+    },
+  },
+}