+++ spec/controllers/api/ui/v1/course_events_controller_spec.rb
@@ -1,11 +1,17 @@
 require "spec_helper"
 
 describe Api::Ui::V1::CourseEventsController, type: :api do
+  before(:all) do
+    ENV['ORGANIZATION']='1'
+  end
+  after(:all) do
+    ENV.delete 'ORGANIZATION'
+  end
+
   before do
     header 'Content-Type', 'application/json; charset=utf-8'
     header 'X-Auth-Token', jwt_token
     header 'Accept-Language', language
-    ENV['ORGANIZATION']='1'
   end
   let(:user) { user_stefania }
   let(:jwt_token) { authenticated_user(user.email) }